+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/profiling/ebpf/storage/EBPFProfilingTargetType.java
@@ -35,6 +35,8 @@ public enum EBPFProfilingTargetType {
     ON_CPU(1),
 
     OFF_CPU(2),
+
+    NETWORK(3),
     ;
     private final int value;
     private static final Map<Integer, EBPFProfilingTargetType> DICTIONARY = new HashMap<>();